+++ packages/client/src/runtime.ts
@@ -1,27 +1,66 @@
-import { Observable, combineLatest, filter, map } from "rxjs"
 import {
   ChainHead$,
   RuntimeContext,
   getObservableClient,
 } from "@polkadot-api/observable-client"
+import {
+  Observable,
+  ReplaySubject,
+  combineLatest,
+  connectable,
+  filter,
+  firstValueFrom,
+  map,
+} from "rxjs"
+import { DescriptorsValue } from "./descriptors"
+
+export enum OpType {
+  Storage = 0,
+  Tx = 1,
+  Event = 2,
+  Error = 3,
+  Const = 4,
+}
 
 export class Runtime {
-  protected _ctx: unknown
-  protected _checksums: string[]
+  private constructor(
+    private _ctx: RuntimeContext,
+    private _checksums: string[],
+    private _descriptors: DescriptorsValue,
+  ) {}
 
-  private constructor(ctx: RuntimeContext, checksums: string[]) {
-    this._ctx = ctx
-    this._checksums = checksums
+  /**
+   * Internal implementation detail. Do not use
+   */
+  static _create(
+    ctx: RuntimeContext,
+    checksums: string[],
+    descriptors: DescriptorsValue,
+  ) {
+    return new Runtime(ctx, checksums, descriptors)
+  }
+
+  /**
+   * Internal implementation detail. Do not use
+   */
+  _getCtx() {
+    return this._ctx
+  }
+
+  /**
+   * Internal implementation detail. Do not use
+   */
+  _getPalletChecksum(opType: OpType, pallet: string, name: string) {
+    return this._checksums[this._descriptors.pallets[pallet][opType][name]]
+  }
+
+  /**
+   * Internal implementation detail. Do not use
+   */
+  _getApiChecksum(name: string, method: string) {
+    return this._checksums[this._descriptors.apis[name][method]]
   }
 }
-export function getRuntimeContext(runtime: Runtime): RuntimeContext {
-  return (runtime as any)._ctx
-}
-function getImportedChecksum(runtime: Runtime, idx: number): string {
-  return (runtime as any)._checksums[idx]
-}
-const createRuntime = (ctx: RuntimeContext, checksums: string[]) =>
-  new (Runtime as any)(ctx, checksums)
 
 export type RuntimeApi = Observable<Runtime> & {
   latest: () => Promise<Runtime>
@@ -29,37 +68,25 @@ export type RuntimeApi = Observable<Runtime> & {
 
 export const getRuntimeApi = (
   checksums: Promise<string[]>,
+  descriptors: Promise<DescriptorsValue>,
   chainHead: ReturnType<ReturnType<typeof getObservableClient>["chainHead$"]>,
 ): RuntimeApi => {
-  let latestRuntime: Promise<Runtime>
-  let resolve: ((r: Runtime) => void) | null = null
-
-  latestRuntime = new Promise<Runtime>((res) => {
-    resolve = res
-  })
-
-  const runtimeWithChecksums$ = combineLatest([chainHead.runtime$, checksums])
-  runtimeWithChecksums$.subscribe(([x, checksums]) => {
-    if (x) {
-      const runtime = createRuntime(x, checksums)
-      if (resolve) {
-        resolve(runtime)
-        resolve = null
-      } else {
-        latestRuntime = Promise.resolve(runtime)
-      }
-    } else if (!resolve) {
-      latestRuntime = new Promise<Runtime>((res) => {
-        resolve = res
-      })
-    }
-  })
+  const runtimeWithChecksums$ = connectable(
+    combineLatest([chainHead.runtime$, checksums, descriptors]).pipe(
+      map(([x, checksums, descriptors]) =>
+        x ? Runtime._create(x, checksums, descriptors) : null,
+      ),
+    ),
+    {
+      connector: () => new ReplaySubject(1),
+    },
+  )
+  runtimeWithChecksums$.connect()
 
   const result = runtimeWithChecksums$.pipe(
-    filter(([x]) => Boolean(x)),
-    map(([x, checksums]) => createRuntime(x!, checksums)),
+    filter((v) => Boolean(v)),
   ) as RuntimeApi
-  result.latest = () => latestRuntime
+  result.latest = () => firstValueFrom(result)
 
   return result
 }
@@ -70,12 +97,13 @@ export interface IsCompatible {
 }
 
 export const compatibilityHelper =
-  (runtimeApi: RuntimeApi, checksumIdx: number) =>
+  (
+    runtimeApi: RuntimeApi,
+    getDescriptorChecksum: (runtime: Runtime) => string,
+  ) =>
   (getChecksum: (ctx: RuntimeContext) => string | null) => {
     function isCompatibleSync(runtime: Runtime) {
-      const ctx = getRuntimeContext(runtime)
-      const checksum = getImportedChecksum(runtime, checksumIdx)
-      return getChecksum(ctx) === checksum
+      return getChecksum(runtime._getCtx()) === getDescriptorChecksum(runtime)
     }
 
     const isCompatible: IsCompatible = (runtime?: Runtime): any => {
@@ -88,7 +116,7 @@ export const compatibilityHelper =
     const waitChecksums = async () => {
       const runtime = await runtimeApi.latest()
       return (ctx: RuntimeContext) =>
-        getChecksum(ctx) === getImportedChecksum(runtime, checksumIdx)
+        getChecksum(ctx) === getDescriptorChecksum(runtime)
     }
     const compatibleRuntime$ = (
       chainHead: ChainHead$,